augur(frontend): split Product workspace into its own module
The Product page lived inline in app.tsx while Budget and Calibration each had their own workspace
module. Extract it to product.tsx (mirroring budget.tsx / calibration.tsx), and move the shared
header chrome it needs — the tab bar, deployment-status chip, and tab routing — into header.tsx so
app.tsx, product.tsx, and the Budget/Calibration surfaces all import it without a cycle. app.tsx is
now just the shell: bootstrap fetch, cross-tab shared-control state, and surface selection.

Pure move — no behavior change (the visual goldens render byte-identically). product.tsx added to
tailwind_content_srcs so its classes aren't purged from the stylesheet.

BAZEL_TEST_INVOCATIONS=buildbuddy:5aabc0e9-bbf3-4fd0-a48d-621b325a6b26
